social_media_analysis/requirements.txt
praw>=7.7
requests>=2.31
tqdm>=4.66
python-dotenv>=1.0
